pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zulip-mcp"
version = "0.1.0"
description = "MCP server for summarizing Zulip channel activity"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "zulip>=0.9.0",
    "pyyaml>=6.0",
]

[project.scripts]
zulip-mcp = "zulip_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/zulip_mcp"]
